annotate allowlist.go refactor filename
